Fix unit and instrumentation tests
