+++ datacapture/src/androidTest/java/com/google/android/fhir/datacapture/views/AttachmentViewHolderFactoryEspressoTest.kt
@@ -28,6 +28,7 @@ import androidx.test.platform.app.InstrumentationRegistry
 import com.google.android.fhir.datacapture.R
 import com.google.android.fhir.datacapture.TestActivity
 import com.google.android.fhir.datacapture.validation.NotValidated
+import com.google.android.fhir.datacapture.views.factories.AttachmentViewHolderFactory
 import com.google.common.truth.Truth.assertThat
 import org.hl7.fhir.r4.model.Attachment
 import org.hl7.fhir.r4.model.CodeType
